perl 3.0 patch #15 (combined patch)
In patch 13, there was a fix to make the VAR=value construct
in a command force interpretation by the shell.  This was botched,
causing an argv list to be occasionally allocated with too small
a size.  This problem is hidden on some machines because of
BSD malloc's semantics.

The lib/dumpvar.pl file was missing final 1; which made it
difficult to tell if it loaded right.

The lib/termcap.pl Tgetent subroutine didn't interpret ^x right
due to a missing ord().

In the section of the man page that gives hints for C programmers,
it falsely declared that you can't subscript array values.  As of
patch 13, this statement is "inoperative".

The t/op.sleep test assumed that a sleep of 2 seconds would always
return a value of 2 seconds slept.  Depending on the load and
the whimsey of the scheduler, it could actually sleep longer than
2 seconds upon occasion.  It now allows sleeps of up to 10 seconds.