perl 3.0 patch #37 (combined patch)
I tried to take the strlen of an integer on systems without wait4()
or waitpid().  For some reason this didn't work too well...

In hash.c there was a call to dbm_nextkey() which needed to be
ifdefed on old dbm systems.

A pattern such as /foo.*bar$/ was wrongly optimized to do
tail matching on "foo".  This was a longstanding bug that
was unmasked by patch 36.

Some systems have some SYS V IPC but not all of it.  Configure
now figures this out.

Patch 36 put the user's PATH in front of Configures, but to make
it work right I needed to change all calls of loc to ./loc in
Configure.

$cryptlib needed to be mentioned in the Makefile.

Apollo 10.3 and Sun 3.5 have some compilation problems, so I
mentioned them in README.

Cray has weird restrictions on setjmp locations--you can't say
if (result = setjmp(...))

Random typos and cleanup.