perl 3.0 patch #42 (combined patch)
Most of these patches are pretty self-explanatory.  Much of this
is random cleanup in preparation for version 4.0, so I won't talk
about it here.  A couple of things should be noted, however.

First, there's a new -0 option that allows you to specify (in octal)
the initial value of $/, the record separator.  It's primarily
intended for use with versions of find that support -print0 to
delimit filenames with nulls, but it's more general than that:

null
^A
default
CR
paragraph mode
file slurp mode

This feature is so new that it didn't even make it into the book.

The other major item is that different patchlevels of perl can
now coexist in your bin directory.  The names "perl" and "taintperl"
are just links to "perl3.044" and "tperl3.044".  This has several
benefits.  The perl3.044 invokes the corresponding tperl3.044 rather
than taintperl, so it always runs the correct version.  Second, you can
"freeze" a script by putting a #! line referring to a version that
it is known to work with.  Third, you can put a new version out
there to try out before making it the default perl.  Lastly, it
sells more disk drives.   :-)

Barring catastrophe, this will likely be the last patch before
version 4.0 comes out.