perl5.000 patch.0f: [enable metaconfig (PL48) users to regenerate Configure]
This patch enables metaconfig (PL48) users to regenerate Configure.  A few
of the units are considerably enhanced from those distributed with
perl5.000.  (This patch assumes that you do *not* have the units from
perl5beta3i installed.  Sorry if that's an inconvenience.)

In particular, *dbm testing also checks for an appropriate
library function, not just the right header files.

The dynamic loading unit U/dlsrc.U has also been updated to support
SVR4 dynamic loading.

The U/i_db.U unit now checks for the types for the various structure
members that have been causing problems.

Note that at a minimum, you need to be sure you have
the cext='.xs' line from this .package file in your own.  That tells
metaconfig that it should look in .xs files (as well as the usual
.[chy] files) for C symbols.