perl 5.002beta1h patch: pod/Makefile
Include -I../lib so that pod2* can find the appropriate libraries.

The pod names are once again sorted.

The PERL line is wrong.  It should read 
This file is automatically generated, but I happened to do it on
a system without miniperl avaialable, so my script fell back on
the perl default.