PATCH: AutoSplit
Below is a patch which moves where AutoSplit splits a file. Currently AutoSplit
split when it sees a sub. This inhibits individual routines from having their
own use or require statements.

This patch causes AutoSplit to split at the last line starting with
a } before a sub.

The reason behind this is that I have a package which is a collection
of utility routines each requiring separate packages, as I am trying
to reduce what is loaded by autoloading the routines, it seems only
sensible to only 'use' packages when the routine which needs them it
loaded

p5p-msgid: <9603111010.AA29935@tiuk.ti.com>