perl 5.002_01: [patch re-organisation and introduction]
This is my patch _01 to perl 5.002.

The bulk of this patch consists of VMS and OS/2 updates, and the
conversion to use EMBED as the default when building Perl.  It
also includes miscellaneous bugfixes, corrections to the pods,
and minor additions, notably the new $^E and $^O variables.

[editor's note: the changes between this and the perl-5.002_01 tag are
the individual changes in the patch set]