Minor potential bug in AV creation
Fix more spots where we had PerlIO_stderr() and should have had
Perl_debug_log instead.

I wasn't the one who originated this patch.  But it looks like it
would improve the safety of AV creation.

Remove potentially incorrect casts on PerlIO_set_ptrcnt.
'ptr' is already STDCHAR, which is supposed to be the type of
char used in stdio.h, so we shouldn't have to cast it.