Re: truncate with file name does not work (with patch)
The prototype for truncate was changed so that perl won't die
with C<use strict;> when the first arg is a bareword (filehandle).
I think it was Tom (as in "tchrist") who brought this up.

Here's a patch that undoes the damage, makes it work with
C<use strict;>, and adds to the testsuite.