perl 5.003_04: perlio.c
Removes an incorrect prototype for setlinebuf from
perlio.c because it conflicts with the correct declaration in
MachTen's stdio.h (and possibly other stdio's as well).

Secondly, the code in perlio.c is not handling the (!PERLIO_IS_STDIO &
HAS_F[GS]ETPOS) case.  The patch fixes this omission (in a rather lumpen
manner).  I don't think this should affect platforms which try to hack a
different path through the #ifdef forest, but this assertion would benefit
from testing...

Dominic Dunlop