Perl 5.003.03: race condition in t/lib/io_sock.t
io_sock.t works by forking a subprocess it can communicate with.
It has the subprocess wait for the main process by sleeping 10
seconds or until an alarm arrives.

With my setup, the alarm signal arrives *before* the child
has a chance to ignore the alarm signal.

I fixed this by moving the "$SIG{ALRM} = sub {};" up before the
fork. It does not hurt to have the parent ignore alarms, too.