perl 5.003_05: lib/AnyDBM_File.pm
AnyDBM_File (modifying ISA does not work as expected)
Now behaves as documented:  Modifying ISA works to select
order in which *DB* modules are tried.  The default is still
the same.

Add helpful "die" message to end of AnyDBM_File.  Previously
it would return a 0, and the failure would eventually show up
somewhere else in the script and be hard to track down.  It is
a failure if perl can't open AnyDBM_File.  The test regression
suite is supposed to indicate this as a failure too.