perl 5.003_06: toke.c
Date: Sat, 14 Sep 1996 17:08:16 -0400
From: Gurusamy Sarathy <gsar@engin.umich.edu>
Subject: whitespace induced lexer errors (with patch)

I finally got around to fixing skipspace() to not indiscriminately
overwrite oldbufptr and oldoldbufptr (which are used in making
expectation decisions in the lexer).

Date: Sat, 14 Sep 1996 18:55:16 -0400
From: Gurusamy Sarathy <gsar@engin.umich.edu>
Subject: perl lexer won't accept C<my($a,$b);$a<=>$b;>

Date: Thu, 19 Sep 1996 11:58:22 -0400
From: "Randy J. Ray" <rjray@uswest.com>
Subject: Patch: Untaint FH flag and clean DATA handles

This patch adds a IOf_UNTAINT flag in sv.h, as one of the possibles
on an xpvio->xio_flags struct member. It is used to mark the given
file handle as a clean source, even when tainting is turned on.
There are also patches to pp_sys.c in pp_sysread to check this flag
before tainting data, and in pp_hot.c in do_readline for the same
reason. Lastly, it patches toke.c to automatically set this flag on
on a __DATA__ filehandle. The creation of the $pack::DATA
pseudo-filehandle is already checked against running under eval, so
this should not introduce any insecurity.

This patch *does not*:

* Add the "untaint" keyword.