Pod typos, pod2man bugs, and miscellaneous installation comments
Here is a patch for various typos and other defects in the Perl
5.003_05 pods, including the pods embedded in library modules.

Bugs found in pod2man

The following bugs were noticed, and some fixed:

1. Where a L<> link extends over more than one line, pod2man does not
treat it as a link but displays it literally, and so these have been
rearranged to place the link on one line.  This is the only bug worked
around.  [Fixed; the rearrangements, which were done beforehand,
remain in some cases, but are no longer necessary, and pod paragraphs
can now be safely reformatted to whatever width is desired in the pod,
without breaking links.]

2. It seems to swallow spaces after certain links: for example, part
of the "open" entry in the perlfunc manpage comes out as "the
\f(CWbinmode\fR entry elsewhere in this documentfor tips", the source
having been "L</binmode> for tips".  [Fixed.]

3. 'L</"Pass by Reference">', in perlsub.pod, comes out as '\fI/"Pass
by Reference\fR', that is, with an initial '/"'.

4. If a pod line begins with ".", nothing is done to prevent [tng]roff
from treating it as a [tng]roff instruction.

5. When the paragraph below =head1 NAME has more than one line, this
confuses pod2man: so in the case of Term::Readline, the manpage begins
with a stray line 'no real package is found, substitutes stubs instead
of basic functions."'.

Of course, it would be better to fix pod2man; I hope that the new Pod
modules, when ready, will not have these defects.