perl 5.003_06: pp_sys.c
Restore the 5.003 gv_fullname() and gv_efullname() functions.
Provide new 3-arg forms gv_fullname3() and gv_efullname3().

Date: Thu, 19 Sep 1996 11:58:22 -0400
From: "Randy J. Ray" <rjray@uswest.com>
Subject: Patch: Untaint FH flag and clean DATA handles

This patch adds a IOf_UNTAINT flag in sv.h, as one of the possibles
on an xpvio->xio_flags struct member. It is used to mark the given
file handle as a clean source, even when tainting is turned on.
There are also patches to pp_sys.c in pp_sysread to check this flag
before tainting data, and in pp_hot.c in do_readline for the same
reason. Lastly, it patches toke.c to automatically set this flag on
on a __DATA__ filehandle. The creation of the $pack::DATA
pseudo-filehandle is already checked against running under eval, so
this should not introduce any insecurity.

This patch *does not*:

* Add the "untaint" keyword.

Date: Sun, 22 Sep 1996 17:26:57 -0400
From: "Randy J. Ray" <rjray@uswest.com>
Subject: Patch to patch for untainting

The following patch ensures that a glob used as a filehandle that
has had the UNTAINT flag set will not carry that flag over on a
re-open. In a nutshell, a re-open of the DATA filehandle would be
considered untainted, and an object of class IO::Handle (or one of
its sub-classes) that is marked untainted with the untaint method,
then closed and re-opened, retained the untaintedness.

Date: Mon, 30 Sep 1996 00:54:37 -0400
From: Spider Boardman <spider@Orb.Nashua.NH.US>

First, with IO::untaint, the patches as posted resulted in a
miniperl which couldn't open files, so the autosplitting of the
library and the creation of Makefiles for the extensions didn't
work.  Worse, it didn't just fail to open files, it dumped core.