perl 5.003_07: t/lib/anydbm.t t/lib/db-btree.t t/lib/db-hash.t t/lib/db-recno.t t/lib/gdbm.t t/lib/ndbm.t t/lib/odbm.t t/lib/sdbm.t
Date: Wed, 9 Oct 1996 22:30:38 -0400 (EDT)
From: Ilya Zakharevich <ilya@math.ohio-state.edu>

File mode under OS/2 is not what you expect. However, this has
nothing to do with databases, _and_ there is a test
for this in stat.t (which dutifully fails). There is
no point to consider this behaviour as a bug in
database code.
So OS/2 is special-cased in these tests.