Re: db-recno.t failures with _11 on Freebsd 2.1-stable
The problem with the new db-recno.t bval tests on 2.1 Freebsd systems is
that Freebsd contains an old version of the DB library (1.74 on
2.1-stable as of early November).  The recno bval handling in this
version of the library is simply broken (bval is ignored).

Here's some documentation.  We're probably going to get a lot of bug
reports about this; I wish there was a better answer.  I don't think I'd
want to call disabling the tests on systems with the old library better,
though.

p5p-msgid: <pzohg0r5tr.fsf@eeyore.ibcinc.com>