flock() constants
Could we have the flock() constants in a standard module?  Since we're
supporting flock() via emulation there are lots of systems in which
these constants can't be pulled in via h2ph.

I don't much care where the constants are stored, so long as they're
somewhere.  Here's a patch which just stuffs them in Fcntl.  The @EXPORT
list is the same as before, you have to ask for these constants
explicitly (via ':flock').

p5p-msgid: <26669.850977437@eeyore.ibcinc.com>