perllocale.pod -- second draft
My notes on this are in a second mailing in this thread.  Please read them
before you respond to this mail.  Thanks.

[editor's note: he is probably referring to his first draft,
<v03007809aedafbad79e9@[194.51.248.70]>, notes below]

Subject: Draft perllocale.pod  -- proposed as replacement for perli18n.pod

Herewith a draft of perllocale.pod.  It's based on Chip's perl18n.pod,
but beefed up considerably, and rearranged a bit.

I'd like to see the name changed, as "i18n" sounds too buzzy to me,
and there was a discussion on p5p some months back which I thought
ended up with the same view.  (Chapter and verse can be supplied if
you want.)  But if consensus (or expedience) is now for perli18n, I
shan't shed more than a few tears.

If consensus is that this pod is close enought to being ready for
prime time for inclusion in 5.004, I'll undertake to munge it in
response to comments, and to fix up all the necessary
cross-referencing in other pods (there's quite a lot of this) by the
end of this week.

If consensus is that this pod can't be made good enough soon enough
(or may never be good enough), I'll adopt a more relaxed timetable (or
none at all): I wouldn't want to hold things up.

May I ask as many people as possible to scrutinize the spelling,
English, mark up and so.  And to think about the points in the
editor's notes.  And PLEASE to try the examples on your own hosts.
Thanks.

p5p-msgid: <v03007800aee1923e30a2@[194.51.248.68]>