Fixes for Interactive Unix
This is a bug report for perl from win@in.rhein-main.de,
generated with the help of perlbug 1.14 running under perl 5.00313.

With the following patches, perl compiles and passes all tests on
Interactiv Unix Version 4.1. I tested with cc and with gcc.

p5p-msgid: <m0vbeNO-00003WC@incom.rhein-main.de>