perlpod.pod possible patches
I am discouraged because I have pretty much determined that a current
incompatibility between Mail::Mailer, perl 5.003_19, and perlbug has
resulted in a number of patchs and bug reports that I have sent over the
past week appear to have disappeared into a sink hole.  Anyways, I will
attempt to do this by hand.  The following is a patch to perlpod.pod to
fix some things that are not formatting properly or which appear to be
incomplete thoughts.

p5p-msgid: <9701070756.AA1185@cas.org>
Signed-off-by: x2487 <lvirden@cas.org>