Misc. doc patches missing in _20
Subject: doc patches missing in _20

Here are a couple of the doc patches I sent to the list which didn't get
into _20.  The first looks like an oversight (a related hunk got in),
but perhaps the second was left out intentionally, in anticipation of
changes in signal restart handling?

These are re-diffed against _20.

p5p-msgid: <102.852695733@eeyore.ibcinc.com>