Fix for anon-lists with tied entries coredump
[George Hartlieb, a MLDBM user reported this problem in private mail.]

The following hypothetical construct:

   for $k (keys %o) {
       foo([$o{$k}]);
   }

coredumps reliably when %o is a tied hash and the FETCH for the
value $o{$k} is substantial enough to cause a stack reallocation.

Patch against 3_19 attached.

 - Sarathy.
   gsar@engin.umich.edu

P.S: Whatever happened to the stack-of-stacks patch?  Even the first
version of that patch would have eliminated this problem.  There may
be many more places where such a fix may be necessary--it's impossible
to find them all.  Please, let's atleast include a #ifdef-ed version
of that patch!

p5p-msgid: <199701100745.CAA13057@aatma.engin.umich.edu>