Advice on TEST failure
At 18:00 +0100 1997-01-09, Paul Marquess wrote:
>What are peoples views on this one. Should I simply write a diagnostic
>message to STDERR when test 51 fails? Some thing like this...
>
> base/cond.........ok
> lib/db-btree......ok
> lib/db-hash.......ok
> lib/db-recno...... not ok 51
>    There is a known problem with older versions of Berkeley DB. You
>    can ignore the errors if you're never going to use the broken
>    functionality (recno databases with a modified bval), otherwise
>    you'll have to upgrade your DB library or OS.  If you feel like
>    upgrading, the most recent version is 1.85. Check out
>    http://www.bostic.com/db for more details.

Looks good to me.  As of 5.003_20, hints/broken-db.msg is a shell scriptlet
that contains a similar (but not quite as useful) message for use by
Configure.  It Would Be Nice if the same text (the same file?) were used in
both situations.  But no big deal.

And, shortly afterwards, Tom Phoenix wrote:
>You're right. Who needs fine print? :-)  What if 'make test' wrapped up
>with a message like this one?
>
>   Failed 1 test, 96.55% okay.
>   u=1.13  s=3.62  cu=60.49  cs=73.18  files=140  tests=3165
>
>   ### Since not all tests were successful, you may want to run
>   ### some tests individually to see the diagnostic messages,
>   ### if any. See the INSTALL document's section on "make test".

That's a good idea, too, whether or not lib/db-recno blats something onto
the screen even when TEST is run without -v.  Here's a patch which
implements it.  (Well, with slightly different ordering.)  I've also taken
the liberty of removing the RCS information from TEST, as it's meaningless
by now, of making the exit behaviour the same for one and for multiple
failures, and of trying to make the difference between "files" and "tests"
clearer.  (This always confused me: I thought maybe "files" was the number
of files that got touched in the course of the tests.)

p5p-msgid: <v03010d01aefbaefcf3bc@[194.51.248.78]>