Re: MakeMaker and 'make uninstall'
>>>>> Hugo van der Sanden writes:

  > I tried uninstalling a package today, the first time I've done so. The
  > package was MailFolder0.03, and as far as I can see the uninstall:
  > target in the Makefile was provided by MakeMaker. This does:

[...]
 
  > and FULLEXT is defined as 'Mail'. The result was that most of the
  > MailTools package was also removed.

Thanks for the report and _Sorry_.

Folks, I think, I have to deprecate the uninstall target. It was never
documented, it was never really clever, it just caused me a lot of
grief. I'd rather design a new solution from scratch.

If nobody objects, I'd recommend the following patch to
lib/ExtUtils/MM_Unix.pm. It makes use of the "nonono" switch of the
uninstall routine and will not uninstall anything.

Chip, can you sneak that in, please? Thanks!

p5p-msgid: <199701101243.NAA26400@anna.in-berlin.de>