Irix 6.3 & 6.4 and perl5.003_20
>>>>> "John" == John Stoffel <jfs@fluent.com> writes:

John> With a little bit of work I've gotten perl5.003_20 to compile
John> and pass all tests under Irix 6.4 on an Origin2000 box.  The
John> trick is to realize that 'nm' output can't be parsed on this
John> machine.  Here's the hints file to use for _21 when it's
John> released.

Chip et al,

I've also gotten it to compile and pass all tests under Irix 6.3 on an
O2, but I had to make some additions.  I've put both hints files
below, which are the same except for the comments really, but in case
they need any changes, they can be made pretty easily now.

p5p-msgid: <199701132242.RAA14601@jfs.Fluent.COM>