Full documentation generation patch
I propose the following patch for perl5.004. It automates the
generation of the documentation and is derived from the pioneer work
of Tom Christiansen, which was sub-documented, unfortunately.

It allows one to run 'roffitall' to generate the full manual and
table of contents, based on existing installed files only.

Next step: automate the maintainance of the roffitall file list, or
generate it automatically... Well, the current version should be OK
for 5.004.