Faster File::Compare
The following optimization speeds up File::Compare with 30% on my
machine (in a test where most files in fact were different).

One could perhaps also optimize the situation where the two files are
the same thing.  Is it portable to compare (stat)[0,1] for the two
files and return 0 if these two numbers are equal?

p5p-msgid: <199702051342.OAA02753@bergen.sn.no>